+++ modules/avm/variables.tf
@@ -90,6 +90,12 @@ variable "email" {
   description = "Email address of the account"
 }
 
+variable "iam_activity_sns_topic_arn" {
+  type        = string
+  default     = null
+  description = "SNS Topic that should receive captured IAM access events"
+}
+
 variable "kms_key_id" {
   type        = string
   default     = null
@@ -116,18 +122,6 @@ variable "provisioned_product_name" {
   type        = string
   default     = null
   description = "A custom name for the provisioned product"
-}
-
-variable "monitor_iam_access" {
-  type = object({
-    event_bus_arn = string
-    identities = list(object({
-      name = string
-      type = string
-    }))
-  })
-  default     = null
-  description = "Object containing list of IAM Identities that should have their access monitored and the EventBridge Event Bus that should receive captured events"
 }
 
 variable "region" {